Add Files via Upload
About the Data

I began by cleaning and enriching the data. I combined several years of Multiple Listing Service (MLS) data (January 1st 2015-November 30th 2017) into one master Excel spreadsheet, and added a few columns that may be pertinent variables as we continued to explore the data and answer questions. Of note, I added the day of week alongside the listed date (ie, Monday, Tuesday, etc). Furthermore, a ratio was added for list price to the actual sale/closed price. *This is the column which I want to examine as my “key to success” – was the home seller able to achieve a sales price very close to, or above, the price the home was listed for? *

*This dataset had 14,356 observations, and 38 variables.*  There were no duplicate rows. Each MLS listing as a unique number, much like a social security number; so duplicates are easy to identify in Excel. 

I frequently went back and forth between Azure ML and R in this experiment. Something that arose as a potential issue, but also added the potential to enrich the data, was that there was a huge amount of categories within a few particular variables; in particular; Listing & Selling Agents, and Listing and Selling Offices. This was a question that begged answering if there was much effect to it; as this is an element a potential home seller has control over. (They can’t change their house economically from a ranch to a colonial, however, you have every ability to pick the right listing agent, and choosing whom you work with or not as the selling/buyers agent.) Within R, I developed four other segmented datasets, so that the Listing/Selling Agents/Offices could be ranked by sales counts, as well as apply a ranking metric of “High” “Medium” and “Low” production. Using the *Join Data* feature in Azure ML to combine them all with a Left Outer Join. With these new variables, *the dataset was now at 14365 observations, and 46 variables.*